Final project notebook (Work in progress)
I converted the original excel file into a csv file but looks like it can't be coverted to DataFrame. The raw excel file then cleaned to keep only the ethnic origin based on the continent and then saved as a CSV file. The CSV file was importd and converted to DataFrame for analysis. The data can be filtered further based on the ML model selection.